Move addperson to global command
