+++ src/test/java/seedu/address/logic/commands/project/AddPersonCommandTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.fail;
+import seedu.address.logic.commands.global.AddPersonCommand;
 import static seedu.address.testutil.TypicalPersons.DESC_A;
 import static seedu.address.testutil.TypicalPersons.DESC_B;
 import static seedu.address.testutil.TypicalPersons.DESC_C;
