+++ src/test/java/seedu/address/logic/parser/AddPersonCommandParsertest.java
@@ -21,7 +21,7 @@ import static seedu.address.testutil.TypicalPersons.DESC_A;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.project.AddPersonCommand;
+import seedu.address.logic.commands.global.AddPersonCommand;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.GitUserName;
 import seedu.address.model.person.Person;
