+++ src/test/java/seedu/address/logic/parser/MainCatalogueParserTest.java
@@ -9,7 +9,7 @@ import static seedu.address.logic.commands.TaskCommandTestUtil.PLAN_MEETING;
 import static seedu.address.logic.commands.TaskCommandTestUtil.TASK_TIME_RANGE_DESC;
 import static seedu.address.logic.commands.TaskCommandTestUtil.VALID_TASK_PROGRESS_HALF;
 import static seedu.address.logic.commands.TeammateTestUtil.VALID_TEAMMATE_GIT_USERNAME_A;
-import seedu.address.logic.commands.project.AddPersonCommand;
+import seedu.address.logic.commands.global.AddPersonCommand;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ASCENDING_SORT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TASK_ASSIGNEE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TASK_NAME;
